Change the extension syntax
We will now use:

```
impl protocol for Struct {
    ...
}
```

Before the extension block started with an identifier, basically any
word would fit. This meant that typos on high level blocks issued a very
confusing error messaged about how the syntax for an extend block was
wrong, even when the typo was on a struct or enum.

Another possible advantage is that now there's no different names for
the extension and its type. They will always be the same, hopefully this
will mean less confusion when debugging message definitions.